skill_bartering_system/requirements.txt
# requirements.txt
streamlit==1.28.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pandas==2.1.3
secure-smtplib==0.1.1
plotly==5.17.0
protobuf==4.21.12
Pillow>=10.0.0
